DR-542 Search comm pro
