+++ application/commissioning-profiler/chalicelib/handler.py
@@ -1,0 +1,2 @@
+# getting the error listed when running ./scripts/unit-tests.sh so put this dummy file in.
+# No source for code: '/uec-dos-service-search/application/commissioning-profiler/chalicelib/handler.py'.
