--- application/utilities/chalicelib/readme.txt
@@ -1,3 +1,0 @@
-Directory structure required for unit testing of utilities
-
-Actual utility code lives in ./application-utils
